Fix default camera flipped Y axis
Changing the camera to default camera after drawing
to the render target will flip the Y axis
This is an issue in the core macroquad code where
a camera without a render target has its Y axis
flipped for some reason

This causes issues for any drawing done after
resetting to the default camera
Currently the only things drawn after resetting to
default camera are the chunk borders, because
we want to draw these at native resolution,
not at the render target resolution.

In the future we will most likely draw more
to the default camera, this fix will enable that

How this is fixed:
After everything is drawn to the render target,
Flip the render target texture on the Y axis.
Then draw the render target texture to the default camera
and flip the default camera's Y axis again aswell.
Now the render target will be rendered normally
(becuase it is flipped twice)