General debugging work, mostly for shaders
- Added a toggle shader button to toggle shaders on or off
- Added a read shader button to read shaders during runtime. Very usefull
for writing and debugging shaders
Pressing the button will read the fragment and vertex shader glsl source
files. These will be read as a raw String. the String is saved in the
UserInterfaceData struct. this struct is read during the App update call
to see if the vertex or fragment shader String in the UserInterfaceData
matches the one in App. If it mismatches, it recompiles the shader
program by calling the load_material() macroquad function
- Reorganized the Debug UI
- Fixed bug where hcunks would not redraw if you press 'reset pixelgrid'
  button, because self.updated_last_frame was not set when chunk.clear()
  was being called.