Shader now uses Texture uniform
Instead of relying on _ScreenTexture. The shader now uses Texture
This can be used to read the input Texture, and change colours or add
different textures based on the pixel colours

The idea is to read which pixel type is on the screen by checking the
colour, then if it matches any of the predefined pixel colors like a
Beige color for Sand pixeltype, then change the texture at that screen
position to a sand texture.